requirements.txt
aiogram~=2.22.1
python-dotenv~=0.20.0
psycopg2-binary~=2.9.3
SQLAlchemy~=1.4.40
lxml~=4.9.1
requests~=2.28.1
beautifulsoup4~=4.11.1
